requirements.txt
flask
python-dotenv
torch
transformers
bcrypt
twilio
requests
nltk 
cryptography 
sentence-transformers 
faiss-cpu
deep-translator
google-generativeai
Flask-WTF
email-validator
pandas
scikit-learn
matplotlib
seaborn
xgboost
joblib